feat: implement cart update functionality and enhance cart components

+++ backend/src/main/java/com/omkashyap/com/backend/dto/projection/CartItemProjection.java
@@ -16,7 +16,7 @@ public interface CartItemProjection {
 
   Integer getTotalReviews();
 
-  Float getAverageRating();
+  Double getAverageRating();
 
   Float getPrice();
