+++ backend/src/main/java/com/omkashyap/com/backend/dto/responseDto/CartItemResponseDto.java
@@ -18,7 +18,7 @@ public class CartItemResponseDto {
   private String productThumbnail;
   private Boolean inStock;
   private Integer totalReviews;
-  private Float averageRating;
+  private Double averageRating;
   private Float price;
   private Integer coins;
   private String productUrl;
